Disable caching as it is not working.
The cacheing of the resources should be done smart: The applications
should be able to tell the validity (timelife) of each requested
resources and resources should be persistent. The work flow is like
this: Request a resource, local storage is queries for timestamp of such
resource and if one exists the timestamp is checked upon the validity of
the resource as requested. If it is valid return the cache, else the
resource is queries on the server and cached again for the give time.