Use unsafe clone with Date objects path
The default unsafe clone implementation in goog namespace does not take care of
date objects and copies them as plain objects. This implementation uses the correct type
and thus assures the date data is clones correctly. The implementation is immediately
useful in the listitem datastructure class where raw data is cloned