Initial attamp at abstracting touch gestures.
The intent is to be able to subscribe any component to the touch events 
and allow the agent to schedule when to fire the actual events. The agent
should also provide access to the data needed by the component like
touch duration and distance. Additional work is required to actually detect
gestures.