Utilize the actual last registered point when calculating velocities.
Because we do not have control over the RAF fireing it is possible to
have the last recorded movement point and the current (at the time of
the touchend) points to be the same and thus we produce velocity of 0
which is incorrect. This solved the issue by holding reference on the
last touch position that is not touched by the RAF handler.