Expose some of the event handling methods.
This allows the user to re-route events from one object to another.
Also fixes a typo in the class name. Scope should be used to avoid this 
kind of bugs.

TODO: pros and cons for letting each control to handle the same event versus
exposing method to control the behavior.