Makes the ripple opacity configurable
Makes the ripple act more in sync with the original paper design in that is animates to a full wave state while pressed and stops there and fades out only on release. Also now is supported configuration in runtime on the opacity of the wave. Wave now does not use the pointer agent by default to avoid unnesesary listeners as it is mostly used as an effect on other elements, but still its ability to be used stand alone is preserved