Add development mode icon loading
Implements alternative icon loading mechanizm that works with base xml files to ease and speed up icon development by generating the renderers on runtime as suggested in https://github.com/pstjvn/pstj-closure/issues/9#issuecomment-62300664 - this now makes it possible to test and develop without producing the templates and classes needed for compilation. The code needed for compilation should be made possible to be generated later. For now for compilation the code needs to be migrated manually and all the renderer classes created