Alternates how we render and decorate elements.
This is an attempt to resolve the decorate internal stage problem. 

In fact all that the decoration is doing must be possible to be done imperatively, so some of the elements will need more work to get in line with this new rule, but it will allow us to compile complete apps with HTML only and thus render much faster on low end devices. It will also make us compatible with the rest of the control/component ecosystem for closure.