Fox for material design waves.
On Android the native wave is triggered on elements even when the wave
element was not pressed. This allows for retargeting the click point
when it hits the wave and also allows the same behavior on the web as it
is observed on native.

Example is illustrated in the updated demo for radio buttons.