Slight change to the API - we prefer to be able to reuse instances.
Basically make the creation part of the object so we can chose when to
reuse the same object and load new data in it and when to discard it and
instead use a new instance. Having the creation part of the object
itself also allows for creation of inherent objects that reuse existing 
logic for simpler objects (thus allowing for inheritance and not only 
composition).

We also define how to notify for updates.