Fix issue with compiled mode icons.
The icon used to be added in decoration stage (too early) and it works in
source mode because icons are loaded via XHR (with delay). However in
compiled mode the icons are readily available and the must be added
after the container itslef has been created as a DOM node.