Fix drawer panel UI when dragging the drawer
This fixes the dragging only for touch devices, dragging with the mouse
remains the same (i.e. no overlaying the main panel) as replacing the
element under the mouse while holding it pressed cuases selection
in the nested elements. The dragging is atypical for mouse user so we
let this one as it is and fix it for touch users who tend to use
swipes more often