Fixes material input animation for state set while the input was not visible.
In some cases state is set while the input is not in the DOM tree because
some parent DOM element has display style set to 'none'.

In such case the measurements taken by the code to calculate the
transformation matrix are wrong (size returns 0) and no transformation is
applyed. This leads to state not being properly removed (as it is bound
to the transition end event) and basically it broke the UI of the input.

The naive approach is to bind the state to a timeout assuming the
animation duration.

This fix uses the proper approach - satate is set directly if no proper
transformation matrx can be calculated. The first time the matrix is calculated
properly the 'previous' UI state is set again and only then transitioning is
allowed and the new state set.