Rework icons
New setup allows only icons that are required in the app to be
included in the build thus significantly lowering the payload
in case of big icon collections that use only a small subset.

The change also starts using es6 in repo and thus the gjslint
was disabled for it. It is expected to be replaced by the
closure linter jar