Force unsealable classes in dev mode for demos
Demos use older code written with goog.defineClass which tries to
seal the class using a wrapper to make it compatible with ES6, however
this prevents calls to goog.base in the constructor function as
the checks are performer for superClass_ on the function itself although
the superClass_ is being added to the wrapper and this the call to
goog.base in the constructor throws. This should be reported.
For now just disable class sealing